+++ lib/internal/readline/emitKeypressEvents.js
@@ -10,6 +10,9 @@ const {
   CSI,
   emitKeys,
 } = require('internal/readline/utils');
+const {
+  kSawKeyPress,
+} = require('internal/readline/interface');
 
 const { clearTimeout, setTimeout } = require('timers');
 const {
@@ -47,7 +50,7 @@ function emitKeypressEvents(stream, iface = {}) {
         clearTimeout(timeoutId);
 
         // This supports characters of length 2.
-        iface._sawKeyPress = charLengthAt(string, 0) === string.length;
+        iface[kSawKeyPress] = charLengthAt(string, 0) === string.length;
         iface.isCompletionEnabled = false;
 
         let length = 0;
